Share the countdown, standings, and share-dialog JS across the player and host bundles
