+++ test/integration/client_bundle_test.go
@@ -6,10 +6,10 @@ import (
 	"testing"
 )
 
-// TestClientBundles pins the esbuild client-bundle pipeline (#721, slice 1):
-// the player shells load the bundled entry points, the bundles are served
-// from the embedded FS, and the one cross-tree module (share.js) stays an
-// external runtime import rather than being inlined into the bundle.
+// TestClientBundles pins the esbuild client-bundle pipeline (#721): the player
+// shells load the bundled entry points, the bundles are served from the
+// embedded FS, and the shared share.js dialog module is inlined into the
+// bundle (slice 2) rather than fetched cross-tree from /assets/js/ at runtime.
 func TestClientBundles(t *testing.T) {
 	t.Parallel()
 
@@ -47,16 +47,22 @@ func TestClientBundles(t *testing.T) {
 		})
 	}
 
-	// share.js lives in the web tree (/assets/js/), not the client tree, so
-	// slice 1 keeps it as an external import the browser fetches at runtime.
-	// The bundle must still carry the import statement; if a future change
-	// inlines it, this guard flags that slice-2 work landed early.
-	t.Run("app bundle keeps share.js as an external import", func(t *testing.T) {
+	// share.js moved into the shared source tree (frontend/shared) in slice 2
+	// and is inlined into the client bundle, so the app bundle must NOT carry a
+	// runtime import of /assets/js/share.js any more - the dialog code travels
+	// inside the bundle. The inlined share-dialog markup confirms it is present.
+	t.Run("app bundle inlines share.js instead of importing it cross-tree", func(t *testing.T) {
 		t.Parallel()
 		body := getBody(ctx, t, baseURL+"/client/js/dist/app.js")
-		if want := "/assets/js/share.js"; !strings.Contains(body, want) {
+		if banned := "/assets/js/share.js"; strings.Contains(body, banned) {
 			t.Errorf(
-				"app bundle missing external import %q - share.js must stay a runtime import in slice 1 (#721)",
+				"app bundle still carries the cross-tree import %q - share.js must be inlined in slice 2 (#721)",
+				banned,
+			)
+		}
+		if want := "share-dialog"; !strings.Contains(body, want) {
+			t.Errorf(
+				"app bundle missing inlined share dialog markup %q - share.js should be bundled in (#721)",
 				want,
 			)
 		}
